requirements-cpu.txt
# --- PyTorch (CPU, Py3.11) ---
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0

# --- HF / токенизация (совместимо с Py3.11) ---
huggingface_hub>=0.23
tokenizers>=0.15,<0.20
sentencepiece>=0.1.99
tiktoken>=0.6

# --- OCR (CPU; PaddleOCR 3.x) ---
paddlepaddle==3.1.0 
paddleocr==3.2.0 
onnxruntime==1.18.1             # CPU (опционально для некоторых узлов)

# --- Изображения / геометрия ---
opencv-python-headless>=4.9,<4.12
Pillow>=10.2
shapely>=2.0
pyclipper>=1.3
rapidfuzz>=3.0

# --- PDF / входы ---
pypdfium2>=4.20
pdfminer.six>=20221105

# --- API / утилиты ---
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
python-multipart>=0.0.9
pydantic>=2.5
loguru>=0.7

# --- Базовые ---
numpy>=1.25,<2

